refactor: Lotto 클래스 생성, 로또 번호 비교 로직 이동
