+++ src/test/java/lotto/PurchaseLottoTest.java
@@ -33,26 +33,6 @@ public class PurchaseLottoTest {
         assertThat(lastWonLottoNumber.contains(number)).isTrue();
     }
 
-    @Test
-    public void countWonNumbers_두개로또번호비교검증() {
-        LottoNumber lottoNumber = new LottoNumber(new LottoNumberGeneratorStrategy() {
-            @Override
-            public List<Integer> generateLottoNumber() {
-                List<Integer> lottoNumber = new ArrayList<Integer>();
-                lottoNumber.add(1);
-                lottoNumber.add(2);
-                lottoNumber.add(3);
-                lottoNumber.add(4);
-                lottoNumber.add(5);
-                lottoNumber.add(6);
-                return lottoNumber;
-            }
-        });
 
-        LastWonLottoNumber lastWonLottoNumber = new LastWonLottoNumber("1,2,3,4,5,6");
-
-        PurchaseLotto purchaseLotto = new PurchaseLotto();
-        assertThat(purchaseLotto.countWonNumbers(lottoNumber.getLottoNumbers(), lastWonLottoNumber.getLastWonLottoNumbers()).wonCount()).isEqualTo(6);
-    }
 
 }
